requirements.txt
python-fasthtml
requests
aiohttp
